+++ python/GafferSceneUI/AttributeTweaksUI.py
@@ -97,10 +97,10 @@ Gaffer.Metadata.registerNode(
 
 			"description",
 			"""
-			Turn on to allow location-specific tweaks to be made to inherited
-			attributes. Attributes will be localised to locations matching the
-			node's filter prior to tweaking. The original inherited attributes
-			will remain untouched.
+			Turn on to allow location-specific tweaks to be made to attributes
+			inherited from ancestors or the scene globals. Attributes will be
+			localised to locations matching the node's filter prior to tweaking.
+			The original inherited attributes will remain untouched.
 			"""
 
 		],