+++ src/GafferScene/ScenePlug.cpp
@@ -45,8 +45,17 @@
 #include "IECore/NullObject.h"
 #include "IECore/StringAlgo.h"
 
+#include "boost/algorithm/string/predicate.hpp"
+
 using namespace Gaffer;
 using namespace GafferScene;
+
+namespace
+{
+
+const std::string g_attributePrefix( "attribute:" );
+
+} // namespace
 
 GAFFER_PLUG_DEFINE_TYPE( ScenePlug );
 
@@ -430,7 +439,7 @@ IECore::ConstCompoundObjectPtr ScenePlug::attributes( const ScenePath &scenePath
 	return attributesPlug()->getValue();
 }
 
-IECore::CompoundObjectPtr ScenePlug::fullAttributes( const ScenePath &scenePath ) const
+IECore::CompoundObjectPtr ScenePlug::fullAttributes( const ScenePath &scenePath, bool withGlobalAttributes ) const
 {
 	PathScope pathScope( Context::current() );
 
@@ -450,6 +459,21 @@ IECore::CompoundObjectPtr ScenePlug::fullAttributes( const ScenePath &scenePath 
 			}
 		}
 		path.pop_back();
+	}
+
+	if( withGlobalAttributes )
+	{
+		for( const auto &g : globals()->members() )
+		{
+			if( boost::starts_with( g.first.string(), g_attributePrefix ) )
+			{
+				const std::string attributeName = g.first.string().substr( g_attributePrefix.size() );
+				if( resultMembers.find( attributeName ) == resultMembers.end() )
+				{
+					resultMembers.insert( { attributeName, g.second } );
+				}
+			}
+		}
 	}
 
 	return result;
@@ -519,7 +543,7 @@ IECore::MurmurHash ScenePlug::attributesHash( const ScenePath &scenePath ) const
 	return attributesPlug()->hash();
 }
 
-IECore::MurmurHash ScenePlug::fullAttributesHash( const ScenePath &scenePath ) const
+IECore::MurmurHash ScenePlug::fullAttributesHash( const ScenePath &scenePath, bool withGlobalAttributes ) const
 {
 	PathScope pathScope( Context::current() );
 
@@ -530,6 +554,18 @@ IECore::MurmurHash ScenePlug::fullAttributesHash( const ScenePath &scenePath ) c
 		pathScope.setPath( &path );
 		attributesPlug()->hash( result );
 		path.pop_back();
+	}
+
+	if( withGlobalAttributes )
+	{
+		for( const auto &g : globals()->members() )
+		{
+			if( boost::starts_with( g.first.string(), g_attributePrefix ) )
+			{
+				result.append( g.first.value() );
+				g.second->hash( result );
+			}
+		}
 	}
 
 	return result;